Move sync history workers into workers component
