+++ portal/context/tunnelHistoryContext/index.tsx
@@ -1,37 +1,15 @@
 'use client'
 
-import { useHemi } from 'hooks/useHemi'
-import { useNetworks } from 'hooks/useNetworks'
 import { useSyncHistory } from 'hooks/useSyncHistory'
 import {
-  SyncStatus,
+  HistoryActions,
   type HistoryReducerState,
 } from 'hooks/useSyncHistory/types'
-import dynamic from 'next/dynamic'
-import { createContext, ReactNode } from 'react'
-import { type RemoteChain } from 'types/chain'
+import { createContext, Dispatch, ReactNode } from 'react'
 import {
   type DepositTunnelOperation,
   type WithdrawTunnelOperation,
 } from 'types/tunnel'
-import { useAccount } from 'wagmi'
-
-const SyncHistoryWorker = dynamic(
-  () => import('./syncHistoryWorker').then(mod => mod.SyncHistoryWorker),
-  { ssr: false },
-)
-
-const isReadyOrSyncing = (status: SyncStatus | undefined) =>
-  !!status && ['ready', 'syncing'].includes(status)
-
-const isChainReadyOrSyncing =
-  (history: HistoryReducerState) => (chain: RemoteChain) =>
-    isReadyOrSyncing(
-      history.deposits.find(d => d.chainId === chain.id)?.status,
-    ) ||
-    isReadyOrSyncing(
-      history.withdrawals.find(d => d.chainId === chain.id)?.status,
-    )
 
 type TunnelHistoryContext = {
   addDepositToTunnelHistory: (deposit: DepositTunnelOperation) => void
@@ -39,6 +17,8 @@ type TunnelHistoryContext = {
     withdrawal: Omit<WithdrawTunnelOperation, 'timestamp'>,
   ) => void
   deposits: DepositTunnelOperation[]
+  dispatch: Dispatch<HistoryActions>
+  history: HistoryReducerState
   resyncHistory: () => void
   syncStatus: HistoryReducerState['status']
   updateDeposit: (
@@ -56,6 +36,8 @@ export const TunnelHistoryContext = createContext<TunnelHistoryContext>({
   addDepositToTunnelHistory: () => undefined,
   addWithdrawalToTunnelHistory: () => undefined,
   deposits: [],
+  dispatch: () => undefined,
+  history: { deposits: [], status: 'idle', withdrawals: [] },
   resyncHistory: () => undefined,
   syncStatus: 'idle',
   updateDeposit: () => undefined,
@@ -68,39 +50,11 @@ type Props = {
 }
 
 export const TunnelHistoryProvider = function ({ children }: Props) {
-  const { address, isConnected } = useAccount()
-
-  const l2ChainId = useHemi().id
-  const { remoteNetworks } = useNetworks()
-
-  const [history, dispatch, context] = useSyncHistory(l2ChainId)
-
-  const historyChainSync = []
-
-  // We need to be ready or syncing to return workers
-  if (isConnected && address && ['ready', 'syncing'].includes(history.status)) {
-    // Add workers for every pair L1-Hemi chain
-    historyChainSync.push(
-      ...remoteNetworks
-        .filter(isChainReadyOrSyncing(history))
-        .map(l1Chain => (
-          <SyncHistoryWorker
-            address={address}
-            dispatch={dispatch}
-            history={history}
-            key={`${l1Chain.id}_${l2ChainId}_${address}`}
-            l1ChainId={l1Chain.id}
-            l2ChainId={l2ChainId}
-          />
-        )),
-    )
-  }
+  const context = useSyncHistory()
 
   return (
     <TunnelHistoryContext.Provider value={context}>
       {children}
-      {/* Sync the transaction history per chain in the background */}
-      {historyChainSync}
     </TunnelHistoryContext.Provider>
   )
 }
